+++ mujoco-rs-derive/Cargo.toml
@@ -1,10 +1,25 @@
 [package]
-name = "mujoco-rs-derive"
-version = "0.1.0"
-edition = "2021"
+name         = "mujoco-rs-derive"
+version      = "0.1.0"
+edition      = "2024"
+license.workspace      = true
+repository.workspace   = true
+rust-version.workspace = true
+description  = "Procedural macros for the mujoco-rs crate."
+readme       = "README.md"
+keywords     = ["MuJoCo", "derive", "proc-macro", "macro"]
+categories   = ["development-tools::procedural-macro-helpers"]
+homepage     = "https://crates.io/crates/mujoco-rs"
+documentation = "https://docs.rs/mujoco-rs-derive/"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
 
 [lib]
-proc_macro = true
+proc-macro = true
 
 [dependencies]
 proc-macro2 = "1.0.106"
